Moving routing_algorithm.hpp and cpp to routing_tests.
