+++ map/framework.cpp
@@ -16,7 +16,6 @@
 #include "routing/index_router.hpp"
 #include "routing/online_absent_fetcher.hpp"
 #include "routing/route.hpp"
-#include "routing/routing_algorithm.hpp"
 #include "routing/routing_helpers.hpp"
 
 #include "routing_common/num_mwm_id.hpp"
