+++ routing/CMakeLists.txt
@@ -90,8 +90,6 @@ set(
   router.hpp
   router_delegate.cpp
   router_delegate.hpp
-  routing_algorithm.cpp
-  routing_algorithm.hpp
   routing_exceptions.hpp
   routing_helpers.cpp
   routing_helpers.hpp
